+++ docs/source/changelog.rst
@@ -12,13 +12,21 @@ Changelog
 ..
     For Help: https://releases.readthedocs.io/en/latest/index.html
 
-* :feature:`45` Pearson Strategy is added to the Distance Approach Module.
+* :feature:`45` Pearson Strategy added to the Distance Approach Module.
 * :support:`45` Pearson Strategy documentation.
+* :support:`50` Updated documentation theme to hudsonthames-sphinx-docs.
 
+* :release:`0.4.1 <2021-04-15>`
+* :feature:`43` OU Model Jurek and OU Model Mudchanatongsuk added to the Stochastic Control Approach Module.
+* :support:`43` OU Model Jurek and OU Model Mudchanatongsuk documentation.
+* :feature:`44` CVine Copula and CVine Copula Strategy added to the Copula Approach Module.
+* :support:`44` CVine Copula and CVine Copula Strategy documentation.
 * :feature:`42` Options to sort pairs by zero-crossings, variance, same industry group added to the Basic Distance Strategy.
 * :support:`42` Updated Basic Distance Strategy documentation.
 * :feature:`40` Vine Copula Partner Selection Approaches added to the Copula Approach Module.
 * :support:`40` Vine Copula Partner Selection Approaches documentation.
+
+* :release:`0.3.1 <2021-02-19>`
 * :support:`38` Removed TensorFlow from requirements and adjusted installation guide.
 
 * :release:`0.3.0 <2021-02-16>`